move functions out of one big helper file into more isolated files
